Harmonize program filter type between RPC and RPC Subscriptions
