+++ packages/rpc-graphql/src/resolvers/program-accounts.ts
@@ -1,5 +1,5 @@
 import { Address } from '@solana/addresses';
-import { Commitment, Slot } from '@solana/rpc-types';
+import { Commitment, GetProgramAccountsDatasizeFilter, GetProgramAccountsMemcmpFilter, Slot } from '@solana/rpc-types';
 import type { GraphQLResolveInfo } from 'graphql';
 
 import { RpcGraphQLContext } from '../context';
@@ -21,8 +21,8 @@ export function resolveProgramAccounts(fieldName?: string) {
              * on the client, the default commitment applied by the server is `"finalized"`.
              */
             commitment?: Commitment;
-            dataSizeFilters?: { dataSize: bigint }[];
-            memcmpFilters?: { bytes: string; encoding: 'base58' | 'base64'; offset: bigint }[];
+            dataSizeFilters?: GetProgramAccountsDatasizeFilter[];
+            memcmpFilters?: GetProgramAccountsMemcmpFilter['memcmp'][];
             /**
              * Prevents accessing stale data by enforcing that the RPC node has processed
              * transactions up to this slot
@@ -34,20 +34,7 @@ export function resolveProgramAccounts(fieldName?: string) {
         info: GraphQLResolveInfo,
     ): Promise<AccountResult[] | null> => {
         const programAddress = fieldName ? parent[fieldName] : args.programAddress;
-        let filters:
-            | (
-                  | {
-                        dataSize: bigint;
-                    }
-                  | {
-                        memcmp: {
-                            bytes: string;
-                            encoding: 'base58' | 'base64';
-                            offset: bigint;
-                        };
-                    }
-              )[]
-            | undefined = [];
+        let filters: (GetProgramAccountsDatasizeFilter | GetProgramAccountsMemcmpFilter)[] | undefined = [];
         if (args.memcmpFilters) {
             filters.concat(
                 args.memcmpFilters.map(memcmpFilter => ({
