+++ packages/rpc-types/src/account-filters.ts
@@ -1,3 +1,5 @@
+import { Base58EncodedBytes, Base64EncodedBytes } from './encoded-bytes';
+
 export type DataSlice = Readonly<{
     /** The number of bytes to return */
     length: number;
@@ -5,22 +7,37 @@ export type DataSlice = Readonly<{
     offset: number;
 }>;
 
+type ProgramNotificationsMemcmpFilterBase58 = Readonly<{
+    /**
+     * The bytes to match, as a base-58 encoded string.
+     *
+     * Data is limited to a maximum of 128 decoded bytes.
+     */
+    bytes: Base58EncodedBytes;
+    /** The encoding to use when decoding the supplied byte string */
+    encoding: 'base58';
+    /** The byte offset into the account data from which to start the comparison */
+    offset: bigint;
+}>;
+
+type ProgramNotificationsMemcmpFilterBase64 = Readonly<{
+    /**
+     * The bytes to match, as a base-64 encoded string.
+     *
+     * Data is limited to a maximum of 128 decoded bytes.
+     */
+    bytes: Base64EncodedBytes;
+    /** The encoding to use when decoding the supplied byte string */
+    encoding: 'base64';
+    /** The byte offset into the account data from which to start the comparison */
+    offset: bigint;
+}>;
+
 export type GetProgramAccountsMemcmpFilter = Readonly<{
     /**
      * This filter matches when the bytes supplied are equal to the account data at the given offset
      */
-    memcmp: Readonly<{
-        /**
-         * The bytes to match, encoded as a string using the specified encoding.
-         *
-         * Data is limited to a maximum of 128 decoded bytes.
-         */
-        bytes: string;
-        /** The encoding to use when decoding the supplied byte string */
-        encoding: 'base58' | 'base64';
-        /** The byte offset into the account data from which to start the comparison */
-        offset: bigint;
-    }>;
+    memcmp: ProgramNotificationsMemcmpFilterBase58 | ProgramNotificationsMemcmpFilterBase64;
 }>;
 
 export type GetProgramAccountsDatasizeFilter = Readonly<{
